// packages/preset-ui/src/shortcuts/types.ts
import type { Accordion } from "./accordion";
import type { Avatar } from "./avatar";
import type { Badge } from "./badge/";
import type { Button } from "./button/";
import type { Card } from "./card/";
import type { Checkbox } from "./checkbox";
import type { Divider } from "./divider";
import type { Input } from "./input";
import type { Kbd } from "./kbd";

export type Components = {
	accordion?: Accordion;
	avatar?: Avatar;
	badge?: Badge;
	button?: Button;
	card?: Card;
	checbox?: Checkbox;
	divider?: Divider;
	input?: Input;
	kbd?: Kbd;
};